reorganises components
