--- app/components/staffView.js
@@ -1,15 +1,0 @@
-'use strict'
-
-var React = require('react')
-var InputView = require('./staff/inputView')
-var Home = React.createClass({
-  render: function () {
-    return (
-        <div className='contentContainer'>
-          <InputView />
-      </div>
-    )
-  }
-})
-
-module.exports = Home
