+++ app/index.js
@@ -22,14 +22,14 @@ function requireAuth(nextState, replace) {
 render((
   <Router history={browserHistory}>
     <Route path='/' component={require('./components/app')}>
-      <IndexRoute component={require('./components/customer/filterableBatchRecordsTable')} />
+      <IndexRoute component={require('./components/customer/homePage')} />
       <Route path="guide" component={require('./components/customer/batchNumberGuide')} />
-      <Route path='staff' component={require('./components/staffView')} onEnter={requireAuth}>
+      <Route path='staff' component={require('./components/staff/staffView')} onEnter={requireAuth}>
         <Route path="batchRecords" component={require('./components/staff/inputView')}>
           <Route path="delete/:id" components={ {form: require('./components/staff/delete/deleteForm')} } />
         </Route>
       </Route>
-      <Route path='*' component={require('./components/notFoundPage')}/>
+      <Route path='*' component={require('./components/common/notFoundPage')}/>
     </Route>
   </Router>
 ), document.getElementById('app'))
